more cleanup
long->int, missed a re-init between parser invokes

+++ src/main/java/com/cliffc/aa/type/TypeRPC.java
@@ -1,7 +1,5 @@
 package com.cliffc.aa.type;
 
-import com.cliffc.aa.util.Ary;
-import com.cliffc.aa.type.Bits;
 import com.cliffc.aa.util.SB;
 
 import java.util.BitSet;
@@ -63,7 +61,7 @@ public class TypeRPC extends Type<TypeRPC> {
     return make(_rpcs.meet( tf._rpcs ));
   }
   
-  public long rpc() { return _rpcs.getbit(); }
+  public int rpc() { return _rpcs.getbit(); }
   public boolean test(int rpc) { return _rpcs.test(rpc); }
   @Override public boolean above_center() { return _rpcs.above_center(); }
   @Override public boolean may_be_con()   { return _rpcs.is_con() || _rpcs.above_center(); }
